updated flow integration metrics to use diffusion based integration
Previously, it was treating the arrows as definite velocities rather than probabilities, resulting in frequently becoming stuck when the lines pointed off the edge of the dataset. Additionally, when stuck, it would oscillate between the two most recent points, creating the impression of going against the flow of time.

The new “flow metrics” of 05f use the more robust diffusion based integration by default. I’m rerunning our benchmarks with this modification.

Signed-off-by: Professor Wug <dev@riddle.press>